refactor(analytics): replace useMetrics with useAnalytics across components
This update refactors the analytics integration by replacing instances of `useMetrics` with `useAnalytics` in various components and tests. The changes include updating imports and modifying the usage of tracking functions to align with the new analytics structure. This refactor aims to streamline analytics tracking and improve code maintainability.

- Updated imports from `useMetrics` to `useAnalytics` in multiple files.
- Adjusted function calls to utilize the new analytics methods.
- Enhanced test coverage to reflect the changes in analytics hooks.

CHANGELOG entry: Refactored analytics integration to use useAnalytics instead of useMetrics.
