+++ app/components/Views/NetworkSelector/NetworkSelector.tsx
@@ -102,7 +102,7 @@ import { isNonEvmChainId } from '../../../core/Multichain/utils';
 import { MultichainNetworkConfiguration } from '@metamask/multichain-network-controller';
 import { useSwitchNetworks } from './useSwitchNetworks';
 import { removeItemFromChainIdList } from '../../../util/metrics/MultichainAPI/networkMetricUtils';
-import { MetaMetrics } from '../../../core/Analytics';
+import { analytics } from '../../../util/analytics/analytics';
 import {
   NETWORK_SELECTOR_SOURCES,
   NetworkSelectorSource,
@@ -884,9 +884,7 @@ const NetworkSelector = () => {
       const { NetworkController } = Engine.context;
       NetworkController.removeNetwork(chainId);
 
-      MetaMetrics.getInstance().addTraitsToUser(
-        removeItemFromChainIdList(chainId),
-      );
+      analytics.identify(removeItemFromChainIdList(chainId));
 
       // set tokenNetworkFilter
       const { PreferencesController } = Engine.context;
