+++ app/components/Views/NetworkSelector/useNetworkSwitch.test.ts
@@ -3,7 +3,7 @@ import {
   NetworkType,
 } from '../../hooks/useNetworksByNamespace/useNetworksByNamespace';
 import { useNetworkSelection } from '../../hooks/useNetworkSelection/useNetworkSelection';
-import { useMetrics } from '../../hooks/useMetrics';
+import { useAnalytics } from '../../hooks/useAnalytics/useAnalytics';
 import Engine from '../../../core/Engine';
 
 // Mock the feature flags
@@ -23,8 +23,8 @@ jest.mock('../../hooks/useNetworkSelection/useNetworkSelection', () => ({
   useNetworkSelection: jest.fn(),
 }));
 
-jest.mock('../../hooks/useMetrics', () => ({
-  useMetrics: jest.fn(),
+jest.mock('../../hooks/useAnalytics/useAnalytics', () => ({
+  useAnalytics: jest.fn(),
 }));
 
 // Mock Engine
@@ -80,7 +80,9 @@ const mockUseNetworksByNamespace =
 const mockUseNetworkSelection = useNetworkSelection as jest.MockedFunction<
   typeof useNetworkSelection
 >;
-const mockUseMetrics = useMetrics as jest.MockedFunction<typeof useMetrics>;
+const mockUseAnalytics = useAnalytics as jest.MockedFunction<
+  typeof useAnalytics
+>;
 
 const mockSelectNetwork = jest.fn();
 const mockTrackEvent = jest.fn();
@@ -108,7 +110,7 @@ describe('useSwitchNetworks Feature Flag Tests', () => {
       customNetworksToReset: [],
     });
 
-    mockUseMetrics.mockReturnValue({
+    mockUseAnalytics.mockReturnValue({
       trackEvent: mockTrackEvent,
       createEventBuilder: mockCreateEventBuilder,
       isEnabled: () => true,
@@ -119,7 +121,7 @@ describe('useSwitchNetworks Feature Flag Tests', () => {
       getDeleteRegulationCreationDate: jest.fn(),
       getDeleteRegulationId: jest.fn(),
       isDataRecorded: jest.fn(),
-      getMetaMetricsId: jest.fn(),
+      getAnalyticsId: jest.fn(),
     });
 
     // Mock the event builder
@@ -164,7 +166,7 @@ describe('useSwitchNetworks Feature Flag Tests', () => {
       // Verify that the hooks are properly initialized
       expect(mockUseNetworksByNamespace).toBeDefined();
       expect(mockUseNetworkSelection).toBeDefined();
-      expect(mockUseMetrics).toBeDefined();
+      expect(mockUseAnalytics).toBeDefined();
     });
 
     it('should have all necessary Engine controllers available', () => {
@@ -183,7 +185,7 @@ describe('useSwitchNetworks Feature Flag Tests', () => {
 
     it('should have proper metrics setup', () => {
       // Verify that metrics are properly set up
-      expect(mockUseMetrics).toBeDefined();
+      expect(mockUseAnalytics).toBeDefined();
       expect(mockTrackEvent).toBeDefined();
       expect(mockCreateEventBuilder).toBeDefined();
     });
